Add wheelSieve method and update .gitignore
Added the wheelSieve method to the list of sieve methods in the primes.cpp file, updated method options in the input.cpp and utilities.cpp files. Additionally, corrected a syntax error in the incrementalSieve method declaration in the primes.h file. Finally, the .gitignore file was updated to ignore 'cmake-build-debug/' directory, to keep system generated files out of the repository.

The wheelSieve method was added to provide a new option to the user for finding prime numbers. This is an optimized version of the ancient sieve algorithm that reduces the number of operations required to find all primes below a given limit. The .gitignore update helps to maintain a clean repository.