"Refactor help message into a separate function"
Moved the display of the help information into its own function 'printHelp' in 'main.h' from 'main.cpp'. This was done to improve the readability and maintability of the code by removing a large block of text from the main function. Now, the function 'printHelp' is called whenever the '-h' flag is provided. A redundant newline in 'input.h' was also removed.