Update test cases and logging command in cmake & input files
The prime numbers used in the test cases have been changed from `2147483648` to `99999989` in the '.github/workflows/cmake.yml' file. The reason for this change is to reduce the testing time and get quicker feedback.

Furthermore, the logging command has been updated in 'input.cpp' to print to `std::cout` rather than `std::clog`. This change was made so the total time elapsed will be clearly visible in the terminal output, rather than getting lost in the log stream.